validate 修正

+++ db/migrate/20100125141248_devise_create_users.rb
@@ -13,13 +13,13 @@ class DeviseCreateUsers < ActiveRecord::Migration[5.2]
       t.integer :birthdate_year, null: false
       t.integer :birthdate_month, null: false
       t.integer :birthdate_day, null: false
-      t.integer :phone_number, null: false, unique: true
+      t.bigint :phone_number, null: false, unique: true
       t.string :address_number, null: false
       t.integer :address_prefecture, null: false, default: 0
       t.string :address_name, null: false
       t.string :address_block, null: false
       t.string :address_building
-      t.integer :address_phone_number
+      t.bigint :address_phone_number
       t.text :introduce
       t.string :encrypted_password, null: false, default: ""
